+ header, .ai, .sketch, Let's Piece I
This is the first and only time I’ll comment on the content commits
I’ll be making to this repo. Just starting a new meditative
illustration project based off of the work of Yoko Ono. Hope whoever is
reading this enjoys.

—

Let’s Piece I

—

“500 Noses are more beautiful than one nose. Even a telephone no. is
more beautiful if 200 people think of the same number at the same time.

a) let 500 people think of the same telephone number at once for a
minute at a set time.
b) let everybody in the city think of the word “yes” at the same time
for 30 seconds. Do it often.
c) make it the whole world thinking all the time.”

1960 spring